backend/requirements.txt
flask==3.0.0
numpy==1.24.3
scikit-learn==1.3.0
pandas==2.0.3
scipy==1.11.1
python-dotenv==1.0.0
gunicorn==21.2.0
transformers==4.35.0
torch==2.1.0
sentencepiece==0.1.99
openai==1.3.0
flask-socketio==5.3.4
flask-cors==4.0.0
python-socketio==5.9.0
langdetect==1.0.9
googletrans==4.0.0rc1
ortools==9.7.2996
googlemaps==4.10.0
